supervisor and permissions are fixed

+++ database/migrations/2026_03_13_102203_create_intern_curriculum_assignment_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
                       ->onDelete('restrict');
 
                 // Match manager_accounts.manager_id type
-                $table->integer('assigned_by');
+                $table->unsignedBigInteger('assigned_by');
                 $table->foreign('assigned_by')
                       ->references('manager_id')
                       ->on('manager_accounts');
