+++ database/migrations/2026_03_13_182036_add_invoice_id_to_transactions_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::table('transactions', function (Blueprint $table) {
             // Add invoice_id column if it doesn't exist
             if (!Schema::hasColumn('transactions', 'invoice_id')) {
-                $table->unsignedBigInteger('invoice_id')->after('id');
+                $table->integer('invoice_id')->after('id');
                 $table->foreign('invoice_id')->references('id')->on('invoices')->onDelete('cascade');
             }
         });
